+++ tests/Bulk/ResponseSetTest.php
@@ -18,6 +18,23 @@ class ResponseSetTest extends BaseTest
 {
     /**
      * @group unit
+     */
+    public function testConstructor(): void
+    {
+        list($responseData, $actions) = $this->_getFixture();
+
+        $responseSet = $this->_createResponseSet($responseData, $actions);
+
+        $this->assertEquals(200, $responseSet->getStatus());
+        $this->assertEquals(12.3, $responseSet->getQueryTime());
+        $this->assertEquals([
+            'url' => 'http://127.0.0.1:9200/_bulk',
+            'http_code' => 200,
+        ], $responseSet->getTransferInfo());
+    }
+
+    /**
+     * @group unit
      * @dataProvider isOkDataProvider
      */
     public function testIsOk(array $responseData, array $actions, bool $expected): void
@@ -129,10 +146,17 @@ class ResponseSetTest extends BaseTest
     {
         $client = $this->createMock(Client::class);
 
+        $response = new Response($responseData, 200);
+        $response->setQueryTime(12.3);
+        $response->setTransferInfo([
+            'url' => 'http://127.0.0.1:9200/_bulk',
+            'http_code' => 200,
+        ]);
+
         $client->expects($this->once())
             ->method('request')
             ->withAnyParameters()
-            ->willReturn(new Response($responseData))
+            ->willReturn($response)
         ;
 
         $bulk = new Bulk($client);